Moving statistics to separate folder + refactoring statistic initialisation to use single oncelock
