+++ pumpkin-cli/src/main.rs
@@ -29,6 +29,7 @@ use pumpkin_lib::proof::ProofLog;
 use pumpkin_lib::results::ProblemSolution;
 use pumpkin_lib::results::SatisfactionResult;
 use pumpkin_lib::results::Solution;
+use pumpkin_lib::statistics::statistic_logging::statistic_logger::configure;
 use pumpkin_lib::termination::TimeBudget;
 use pumpkin_lib::variables::PropositionalVariable;
 use pumpkin_lib::Solver;
@@ -376,7 +377,7 @@ fn configure_logging_unknown() -> std::io::Result<()> {
 }
 
 fn configure_logging_minizinc(verbose: bool, log_statistics: bool) -> std::io::Result<()> {
-    pumpkin_lib::statistics::configure(
+    configure(
         log_statistics,
         "%%%mzn-stat:",
         Some("%%%mzn-stat-end"),
@@ -407,7 +408,7 @@ fn configure_logging_sat(
     omit_timestamp: bool,
     omit_call_site: bool,
 ) -> std::io::Result<()> {
-    pumpkin_lib::statistics::configure(log_statistics, "c STAT", None, None);
+    configure(log_statistics, "c STAT", None, None);
     let level_filter = if verbose {
         LevelFilter::Debug
     } else {
