--- pumpkin-lib/src/basic_types/statistics/mod.rs
@@ -1,3 +1,0 @@
-pub(crate) mod counters;
-pub(crate) mod statistic_logger;
-pub(crate) mod statistic_logging;
