+++ pumpkin-lib/src/statistics/statistic_logging.rs
@@ -1,0 +1,76 @@
+//! Responsible for behaviour related to logging statistics with a specific pre-fix and closing
+//! lines.
+
+pub mod statistic_logger {
+    use std::fmt::Display;
+    use std::sync::OnceLock;
+
+    use convert_case::Case;
+    use convert_case::Casing;
+
+    #[derive(Debug)]
+    pub struct StatisticOptions<'a> {
+        // What is printed before a statistic is printed, the statistics will be printed in the
+        // form `{PREFIX} {VALUE}={NAME}`
+        statistic_prefix: &'a str,
+        // A closing line which is printed after all of the statistics have been printed
+        after_statistics: Option<&'a str>,
+        // Whether statistics should be logged
+        log_statistics: bool,
+        // The casing of the name of the statistic
+        statistics_casing: Option<Case>,
+    }
+
+    static STATISTIC_OPTIONS: OnceLock<StatisticOptions> = OnceLock::new();
+
+    /// Configures the logging of the statistics.
+    ///
+    /// It specifies the (optional) prefix and a closing line (postfix) which
+    /// can be printed after all of the statistics have been logged. Statistics will only be printed
+    /// if `log_statistics` is true.
+    pub fn configure(
+        log_statistics: bool,
+        prefix: &'static str,
+        after: Option<&'static str>,
+        casing: Option<Case>,
+    ) {
+        let _ = STATISTIC_OPTIONS.get_or_init(|| StatisticOptions {
+            statistic_prefix: prefix,
+            after_statistics: after,
+            log_statistics,
+            statistics_casing: casing,
+        });
+    }
+
+    /// Logs the provided statistic with name `name` and value `value`. At the moment it will log in
+    /// the format `STATISTIC_PREFIX NAME=VALUE`.
+    pub fn log_statistic(name: impl Display, value: impl Display) {
+        let statistic_options = STATISTIC_OPTIONS
+            .get()
+            .expect("Expected statistics to be configured before logging");
+        if statistic_options.log_statistics {
+            let name = if let Some(casing) = &statistic_options.statistics_casing {
+                name.to_string().to_case(*casing)
+            } else {
+                name.to_string()
+            };
+            println!("{} {name}={value}", statistic_options.statistic_prefix)
+        }
+    }
+
+    /// Logs the postfix of the statistics (if it has been set).
+    ///
+    /// Certain formats (e.g. the [MiniZinc](https://www.minizinc.org/doc-2.7.6/en/fzn-spec.html#statistics-output)
+    /// output format) require that a block of statistics is followed by a closing line; this
+    /// function outputs this closing line **if** it is configued.
+    pub fn log_statistic_postfix() {
+        let statistic_options = STATISTIC_OPTIONS
+            .get()
+            .expect("Expected statistics to be configured before logging");
+        if statistic_options.log_statistics {
+            if let Some(post_fix) = statistic_options.after_statistics {
+                println!("{post_fix}")
+            }
+        }
+    }
+}
